requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.11.2
python-docx>=0.8.11
